[refactor] hitcount 타입 변경 및 소스 수정

+++ src/main/java/step2/Domain/HitCount.java
@@ -4,9 +4,9 @@ import java.util.Objects;
 
 public class HitCount {
 
-    private int hitCount;
+    private long hitCount;
 
-    public HitCount(int hitCount) {
+    public HitCount(long hitCount) {
         this.hitCount = hitCount;
     }
